frontend/requirements.txt
streamlit>=1.33.0
requests>=2.31.0
Pillow>=10.0.0
